pyproject.toml
[project]
name = "electronics-saling-network"
version = "0.1.0"
description = ""
authors = [
    { name = "osha890", email = "oshavshav@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "celery (>=5.5.3,<6.0.0)",
    "qrcode (>=8.2,<9.0)",
    "pillow (>=11.3.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"
pre-commit = "^4.3.0"
django-debug-toolbar = "^6.0.0"
drf-spectacular = "^0.28.0"
faker = "^37.8.0"

[tool.isort]
profile = "black"
known_first_party = ["esn", "products", "network", "employees"]
